# PERF: significantly speed up nanargmin/max for F arrays via coordinate transform