Process `nanargmin/max` in C-order and do coordinate transform to F-order if needed, providing up to a 5x speedup.
```
asv compare HEAD^ HEAD -s --sort ratio --only-changed
       before           after         ratio
     [cad862cc]       [c94520a7]
     <master>         <macro_cleanup>
-      1.74±0.4ms      1.53±0.02ms     0.88  reduce.Time2DReductions.time_nanargmin('float64', (1000, 1000), 'C', 0) [T470/conda-py3.7-numpy1.16-CCgcc-9-CXXg++-9]
-      1.50±0.5μs      1.31±0.01μs     0.87  reduce.Time1DReductions.time_nanargmin('float32', (1000,)) [T470/conda-py3.7-numpy1.16-CCgcc-9-CXXg++-9]
-       969±200μs          823±4μs     0.85  reduce.Time2DReductions.time_nanargmin('int64', (1000, 1000), 'C', 1) [T470/conda-py3.7-numpy1.16-CCclang-9-CXXclang++-9]
-      1.42±0.2ms      1.20±0.01ms     0.84  reduce.Time2DReductions.time_nanargmin('int64', (1000, 1000), 'C', 1) [T470/conda-py3.7-numpy1.16-CCgcc-7-CXXg++-7]
-       985±100μs          823±6μs     0.84  reduce.Time2DReductions.time_nanargmin('int64', (1000, 1000), 'F', 0) [T470/conda-py3.7-numpy1.16-CCclang-9-CXXclang++-9]
-      1.57±0.2ms       1.27±0.1ms     0.81  reduce.Time2DReductions.time_nanargmin('int64', (1000, 1000), 'F', 1) [T470/conda-py3.7-numpy1.16-CCclang-9-CXXclang++-9]
-      3.58±0.1ms      2.84±0.03ms     0.79  reduce.Time2DReductions.time_nanargmin('float32', (1000, 1000), 'C', 0) [T470/conda-py3.7-numpy1.16-CCclang-9-CXXclang++-9]
-        114±20μs         89.4±2μs     0.78  reduce.Time1DReductions.time_nanargmin('float32', (100000,)) [T470/conda-py3.7-numpy1.16-CCgcc-9-CXXg++-9]
-      1.26±0.1ms          985±7μs     0.78  reduce.Time2DReductions.time_nanargmin('float32', (1000, 1000), 'C', None) [T470/conda-py3.7-numpy1.16-CCgcc-9-CXXg++-9]
-        12.4±2ms      9.66±0.06ms     0.78  reduce.Time1DReductions.time_nanargmin('float32', (10000000,)) [T470/conda-py3.7-numpy1.16-CCgcc-9-CXXg++-9]
-      1.81±0.3ms      1.35±0.01ms     0.74  reduce.Time2DReductions.time_nanargmin('int32', (1000, 1000), 'F', 1) [T470/conda-py3.7-numpy1.16-CCgcc-7-CXXg++-7]
-      1.36±0.1μs         993±20ns     0.73  reduce.Time1DReductions.time_nanargmin('int32', (1000,)) [T470/conda-py3.7-numpy1.16-CCgcc-8-CXXg++-8]
-     1.50±0.07ms      1.07±0.01ms     0.71  reduce.Time2DReductions.time_nanargmin('int64', (1000, 1000), 'C', None) [T470/conda-py3.7-numpy1.16-CCgcc-7-CXXg++-7]
-     1.56±0.07ms      1.10±0.03ms     0.71  reduce.Time2DReductions.time_nanargmin('float32', (1000, 1000), 'C', 1) [T470/conda-py3.7-numpy1.16-CCgcc-8-CXXg++-8]
-       85.8±10μs       60.4±0.5μs     0.70  reduce.Time1DReductions.time_nanargmin('float64', (100000,)) [T470/conda-py3.7-numpy1.16-CCgcc-8-CXXg++-8]
-     2.32±0.03ms       1.63±0.2ms     0.70  reduce.Time2DReductions.time_nanargmin('int64', (1000, 1000), 'C', 0) [T470/conda-py3.7-numpy1.16-CCgcc-8-CXXg++-8]
-      1.85±0.6ms      1.28±0.03ms     0.69  reduce.Time2DReductions.time_nanargmin('float64', (1000, 1000), 'F', 0) [T470/conda-py3.7-numpy1.16-CCgcc-9-CXXg++-9]
-      1.49±0.1μs      1.03±0.04μs     0.69  reduce.Time1DReductions.time_nanargmin('float32', (1000,)) [T470/conda-py3.7-numpy1.16-CCgcc-7-CXXg++-7]
-      2.27±0.6ms      1.56±0.09ms     0.69  reduce.Time2DReductions.time_nanargmin('float64', (1000, 1000), 'F', 1) [T470/conda-py3.7-numpy1.16-CCgcc-9-CXXg++-9]
-      1.45±0.2μs          995±8ns     0.69  reduce.Time1DReductions.time_nanargmin('int64', (1000,)) [T470/conda-py3.7-numpy1.16-CCgcc-9-CXXg++-9]
-        11.1±4ms      7.59±0.04ms     0.69  reduce.Time1DReductions.time_nanargmin('float64', (10000000,)) [T470/conda-py3.7-numpy1.16-CCgcc-9-CXXg++-9]
-      1.53±0.4μs         998±20ns     0.65  reduce.Time1DReductions.time_nanargmin('int32', (1000,)) [T470/conda-py3.7-numpy1.16-CCgcc-7-CXXg++-7]
-      1.54±0.2μs      1.00±0.02μs     0.65  reduce.Time1DReductions.time_nanargmin('int32', (1000,)) [T470/conda-py3.7-numpy1.16-CCgcc-9-CXXg++-9]
-       94.3±10μs         59.7±2μs     0.63  reduce.Time1DReductions.time_nanargmin('int32', (100000,)) [T470/conda-py3.7-numpy1.16-CCgcc-8-CXXg++-8]
-        96.2±8μs         60.8±2μs     0.63  reduce.Time1DReductions.time_nanargmin('float64', (100000,)) [T470/conda-py3.7-numpy1.16-CCgcc-7-CXXg++-7]
-     3.89±0.09ms      2.46±0.01ms     0.63  reduce.Time2DReductions.time_nanargmin('float32', (1000, 1000), 'F', None) [T470/conda-py3.7-numpy1.16-CCclang-9-CXXclang++-9]
-      1.93±0.1ms      1.21±0.01ms     0.63  reduce.Time2DReductions.time_nanargmin('int32', (1000, 1000), 'C', 0) [T470/conda-py3.7-numpy1.16-CCgcc-9-CXXg++-9]
-      1.31±0.4ms        804±100μs     0.62  reduce.Time2DReductions.time_nanargmin('int64', (1000, 1000), 'C', None) [T470/conda-py3.7-numpy1.16-CCgcc-9-CXXg++-9]
-      1.38±0.2ms         838±10μs     0.61  reduce.Time2DReductions.time_nanargmin('float32', (1000, 1000), 'C', 1) [T470/conda-py3.7-numpy1.16-CCgcc-7-CXXg++-7]
-      1.12±0.2ms         679±50μs     0.61  reduce.Time2DReductions.time_nanargmin('int32', (1000, 1000), 'C', None) [T470/conda-py3.7-numpy1.16-CCgcc-9-CXXg++-9]
-      1.64±0.3ms         992±20μs     0.60  reduce.Time2DReductions.time_nanargmin('float32', (1000, 1000), 'C', 0) [T470/conda-py3.7-numpy1.16-CCgcc-7-CXXg++-7]
-        102±10μs       59.8±0.9μs     0.58  reduce.Time1DReductions.time_nanargmin('int64', (100000,)) [T470/conda-py3.7-numpy1.16-CCgcc-9-CXXg++-9]
-      1.77±0.3ms      1.03±0.06ms     0.58  reduce.Time2DReductions.time_nanargmin('float32', (1000, 1000), 'F', 1) [T470/conda-py3.7-numpy1.16-CCgcc-7-CXXg++-7]
-      1.27±0.2ms          730±5μs     0.58  reduce.Time2DReductions.time_nanargmin('int32', (1000, 1000), 'F', 0) [T470/conda-py3.7-numpy1.16-CCgcc-9-CXXg++-9]
-      1.42±0.3ms         796±90μs     0.56  reduce.Time2DReductions.time_nanargmin('int64', (1000, 1000), 'C', None) [T470/conda-py3.7-numpy1.16-CCgcc-8-CXXg++-8]
-      4.48±0.1ms      2.50±0.03ms     0.56  reduce.Time2DReductions.time_nanargmin('float64', (1000, 1000), 'F', None) [T470/conda-py3.7-numpy1.16-CCclang-9-CXXclang++-9]
-      1.51±0.2ms        844±200μs     0.56  reduce.Time2DReductions.time_nanargmin('float32', (1000, 1000), 'F', 0) [T470/conda-py3.7-numpy1.16-CCgcc-7-CXXg++-7]
-      1.23±0.2ms         680±80μs     0.55  reduce.Time2DReductions.time_nanargmin('int32', (1000, 1000), 'C', None) [T470/conda-py3.7-numpy1.16-CCgcc-7-CXXg++-7]
-      1.34±0.3ms        740±300μs     0.55  reduce.Time2DReductions.time_nanargmin('int32', (1000, 1000), 'F', 0) [T470/conda-py3.7-numpy1.16-CCgcc-8-CXXg++-8]
-         109±7μs         60.0±2μs     0.55  reduce.Time1DReductions.time_nanargmin('float32', (100000,)) [T470/conda-py3.7-numpy1.16-CCgcc-7-CXXg++-7]
-      1.83±0.1ms         988±10μs     0.54  reduce.Time2DReductions.time_nanargmin('int32', (1000, 1000), 'C', 0) [T470/conda-py3.7-numpy1.16-CCgcc-8-CXXg++-8]
-        14.2±4ms         7.65±1ms     0.54  reduce.Time1DReductions.time_nanargmin('int64', (10000000,)) [T470/conda-py3.7-numpy1.16-CCgcc-9-CXXg++-9]
-        12.6±1ms      6.77±0.04ms     0.54  reduce.Time1DReductions.time_nanargmin('int32', (10000000,)) [T470/conda-py3.7-numpy1.16-CCgcc-9-CXXg++-9]
-        112±10μs       59.5±0.3μs     0.53  reduce.Time1DReductions.time_nanargmin('int32', (100000,)) [T470/conda-py3.7-numpy1.16-CCgcc-9-CXXg++-9]
-        14.3±1ms       7.64±0.5ms     0.53  reduce.Time1DReductions.time_nanargmin('float64', (10000000,)) [T470/conda-py3.7-numpy1.16-CCgcc-8-CXXg++-8]
-      1.25±0.1ms          665±4μs     0.53  reduce.Time2DReductions.time_nanargmin('float32', (1000, 1000), 'C', None) [T470/conda-py3.7-numpy1.16-CCgcc-7-CXXg++-7]
-        116±20μs         61.3±3μs     0.53  reduce.Time1DReductions.time_nanargmin('int32', (100000,)) [T470/conda-py3.7-numpy1.16-CCgcc-7-CXXg++-7]
-      2.45±0.3ms      1.29±0.02ms     0.53  reduce.Time2DReductions.time_nanargmin('float64', (1000, 1000), 'C', 1) [T470/conda-py3.7-numpy1.16-CCgcc-9-CXXg++-9]
-        13.1±2ms       6.74±0.9ms     0.52  reduce.Time1DReductions.time_nanargmin('int32', (10000000,)) [T470/conda-py3.7-numpy1.16-CCgcc-7-CXXg++-7]
-      1.97±0.2ms         992±20μs     0.50  reduce.Time2DReductions.time_nanargmin('int32', (1000, 1000), 'F', 1) [T470/conda-py3.7-numpy1.16-CCgcc-8-CXXg++-8]
-        126±20μs         59.6±2μs     0.47  reduce.Time1DReductions.time_nanargmin('float64', (100000,)) [T470/conda-py3.7-numpy1.16-CCgcc-9-CXXg++-9]
-        16.4±6ms         7.78±3ms     0.47  reduce.Time1DReductions.time_nanargmin('float64', (10000000,)) [T470/conda-py3.7-numpy1.16-CCgcc-7-CXXg++-7]
-      1.71±0.6ms        793±100μs     0.46  reduce.Time2DReductions.time_nanargmin('float64', (1000, 1000), 'C', None) [T470/conda-py3.7-numpy1.16-CCgcc-7-CXXg++-7]
-      2.29±0.6μs      1.03±0.01μs     0.45  reduce.Time1DReductions.time_nanargmin('float64', (1000,)) [T470/conda-py3.7-numpy1.16-CCgcc-7-CXXg++-7]
-      1.75±0.1ms         680±20μs     0.39  reduce.Time2DReductions.time_nanargmin('int32', (1000, 1000), 'C', None) [T470/conda-py3.7-numpy1.16-CCgcc-8-CXXg++-8]
-      2.96±0.4ms      1.15±0.03ms     0.39  reduce.Time2DReductions.time_nanargmin('float32', (1000, 1000), 'F', None) [T470/conda-py3.7-numpy1.16-CCgcc-8-CXXg++-8]
-        17.9±2ms      6.77±0.05ms     0.38  reduce.Time1DReductions.time_nanargmin('int32', (10000000,)) [T470/conda-py3.7-numpy1.16-CCgcc-8-CXXg++-8]
-      1.91±0.3ms         682±10μs     0.36  reduce.Time2DReductions.time_nanargmin('int32', (1000, 1000), 'F', None) [T470/conda-py3.7-numpy1.16-CCclang-9-CXXclang++-9]
-      2.83±0.4ms          992±8μs     0.35  reduce.Time2DReductions.time_nanargmin('float32', (1000, 1000), 'F', None) [T470/conda-py3.7-numpy1.16-CCgcc-9-CXXg++-9]
-      3.33±0.3ms       1.08±0.3ms     0.32  reduce.Time2DReductions.time_nanargmin('int64', (1000, 1000), 'F', None) [T470/conda-py3.7-numpy1.16-CCgcc-7-CXXg++-7]
-      2.61±0.3ms          818±4μs     0.31  reduce.Time2DReductions.time_nanargmin('int64', (1000, 1000), 'F', None) [T470/conda-py3.7-numpy1.16-CCclang-9-CXXclang++-9]
-      3.91±0.5ms       1.06±0.3ms     0.27  reduce.Time2DReductions.time_nanargmin('int64', (1000, 1000), 'F', None) [T470/conda-py3.7-numpy1.16-CCgcc-9-CXXg++-9]
-      3.03±0.4ms         797±10μs     0.26  reduce.Time2DReductions.time_nanargmin('float64', (1000, 1000), 'F', None) [T470/conda-py3.7-numpy1.16-CCgcc-8-CXXg++-8]
-      2.74±0.4ms          680±6μs     0.25  reduce.Time2DReductions.time_nanargmin('int32', (1000, 1000), 'F', None) [T470/conda-py3.7-numpy1.16-CCgcc-7-CXXg++-7]
-      3.25±0.5ms          793±7μs     0.24  reduce.Time2DReductions.time_nanargmin('float64', (1000, 1000), 'F', None) [T470/conda-py3.7-numpy1.16-CCgcc-9-CXXg++-9]
-      2.78±0.4ms          674±5μs     0.24  reduce.Time2DReductions.time_nanargmin('int32', (1000, 1000), 'F', None) [T470/conda-py3.7-numpy1.16-CCgcc-9-CXXg++-9]
-      3.45±0.9ms         792±10μs     0.23  reduce.Time2DReductions.time_nanargmin('float64', (1000, 1000), 'F', None) [T470/conda-py3.7-numpy1.16-CCgcc-7-CXXg++-7]
-      3.54±0.3ms        805±100μs     0.23  reduce.Time2DReductions.time_nanargmin('int64', (1000, 1000), 'F', None) [T470/conda-py3.7-numpy1.16-CCgcc-8-CXXg++-8]
-      3.00±0.4ms         671±10μs     0.22  reduce.Time2DReductions.time_nanargmin('float32', (1000, 1000), 'F', None) [T470/conda-py3.7-numpy1.16-CCgcc-7-CXXg++-7]
-      3.35±0.4ms        703±300μs     0.21  reduce.Time2DReductions.time_nanargmin('int32', (1000, 1000), 'F', None) [T470/conda-py3.7-numpy1.16-CCgcc-8-CXXg++-8]
       before           after         ratio
     [cad862cc]       [c94520a7]
     <master>         <macro_cleanup>
+        67.7±5μs          109±5μs     1.61  reduce.Time1DReductions.time_nanargmin('float32', (100000,)) [T470/conda-py3.7-numpy1.16-CCgcc-8-CXXg++-8]
+     1.09±0.09μs      1.43±0.03μs     1.31  reduce.Time1DReductions.time_nanargmin('float32', (1000,)) [T470/conda-py3.7-numpy1.16-CCgcc-8-CXXg++-8]
```